+++ weblate/trans/forms.py
@@ -997,9 +997,7 @@ class CommentForm(forms.Form):
     comment = forms.CharField(
         widget=forms.Textarea(attrs={'dir': 'auto'}),
         label=_('New comment'),
-        help_text=_(
-            'You can use Markdown or mention other users by @username.'
-        ),
+        help_text=_('You can use Markdown and mention users by @username.'),
         max_length=1000,
     )